requirements.txt
fastapi
pydantic
python-dotenv
requests
supabase
uvicorn
